:recycle: chore: Move bundler code to seperate files that can cmd-p easily
