+++ bundlers/bundler/src/solc/solcPlugin.ts
@@ -1,131 +1,8 @@
-import { SolidityResolver } from '../types'
-import { Logger } from '../types'
+import type { Bundler } from '../types'
 import { getEtherscanLinks } from '../utils'
-import { ModuleInfo, moduleFactory } from './moduleFactory'
-import { ResolvedConfig } from '@evmts/config'
-import { readFileSync } from 'fs'
-import * as resolve from 'resolve'
-// TODO wrap this in a typesafe version
-// @ts-ignore
-import solc from 'solc'
+import { resolveArtifacts, resolveArtifactsSync } from './resolveArtifacts'
 
-// Compile the Solidity contract and return its ABI and bytecode
-function compileContractSync(
-	filePath: string,
-	basedir: string,
-	config: ResolvedConfig['compiler'],
-): solc.CompiledContract | undefined {
-	const source: string = readFileSync(
-		resolve.sync(filePath, {
-			basedir,
-		}),
-		'utf8',
-	)
-
-	const entryModule = moduleFactory(
-		filePath,
-		source,
-		config.remappings,
-		config.libs,
-	)
-
-	/**
-	 * Known bug!!! This can hit a stack error
-	 * We should make this iterative instead of recursive
-	 */
-	const getAllModulesRecursively = (
-		m = entryModule,
-		modules: Record<string, ModuleInfo> = {},
-	) => {
-		modules[m.id] = m
-		/**
-		 * This could get cached
-		 */
-		for (const dep of m.resolutions) {
-			getAllModulesRecursively(dep, modules)
-		}
-		return modules
-	}
-	const allModules = getAllModulesRecursively()
-
-	const sources = Object.fromEntries(
-		Object.entries(allModules).map(([id, module]) => {
-			return [id, { content: module.code }]
-		}),
-	)
-
-	const input: solc.InputDescription = {
-		language: 'Solidity',
-		sources,
-		settings: {
-			outputSelection: {
-				'*': {
-					'*': ['*'],
-				},
-			},
-		},
-	}
-
-	const output: solc.OutputDescription = JSON.parse(
-		solc.compile(JSON.stringify(input)),
-	)
-
-	const warnings = output?.errors?.filter(
-		({ type }: { type: 'Warning' | 'Error' }) => type === 'Warning',
-	)
-	const isErrors = output?.errors?.length > warnings?.length
-
-	if (isErrors) {
-		console.error('Compilation errors:', output?.errors)
-		throw new Error('Compilation failed')
-	}
-	if (warnings?.length) {
-		console.warn('Compilation warnings:', output?.errors)
-	}
-
-	return output.contracts[entryModule.id]
-}
-
-const resolveArtifactsSync = (
-	solFile: string,
-	basedir: string,
-	logger: Logger,
-	config: ResolvedConfig,
-):
-	| Record<string, { contractName: string; abi: any; bytecode: string }>
-	| undefined => {
-	if (!solFile.endsWith('.sol')) {
-		throw new Error('Not a solidity file')
-	}
-	const contracts = compileContractSync(solFile, basedir, config.compiler)
-
-	if (!contracts) {
-		logger.error(`Compilation failed for ${solFile}`)
-		throw new Error('Compilation failed')
-	}
-
-	return Object.fromEntries(
-		Object.entries(contracts).map(([contractName, contract]) => {
-			const abi = (contract as any).abi
-			const bytecode = (contract as any).evm.bytecode.object
-			return [contractName, { contractName, abi, bytecode }]
-		}),
-	)
-}
-
-const resolveArtifacts = async (
-	solFile: string,
-	basedir: string,
-	logger: Logger,
-	config: ResolvedConfig,
-): Promise<
-	| Record<string, { contractName: string; abi: any; bytecode: string }>
-	| undefined
-> => {
-	return resolveArtifactsSync(solFile, basedir, logger, config)
-}
-
-export const solcModules: SolidityResolver = (config, logger) => {
+export const solcModules: Bundler = (config, logger) => {
 	return {
 		name: solcModules.name,
 		config,
