+++ bundlers/bundler/src/types.ts
@@ -1,0 +1,115 @@
+import type { ResolvedConfig } from "@evmts/config"
+
+export type Bundler = (
+  config: ResolvedConfig,
+  logger: Logger,
+) => {
+  /**
+   * The name of the plugin.
+   */
+  name: string
+  /**
+   * The configuration of the plugin.
+   */
+  config: ResolvedConfig
+  include?: string[]
+  exclude?: string[]
+  /**
+   * Resolves .d.ts representation of the solidity module
+   */
+  resolveDts: (module: string, basedir: string) => Promise<string>
+  /**
+   * Resolves .d.ts representation of the solidity module
+   */
+  resolveDtsSync: (module: string, basedir: string) => string
+  /**
+   * Resolves typescript representation of the solidity module
+   */
+  resolveTsModule: (module: string, basedir: string) => Promise<string>
+  /**
+   * Resolves typescript representation of the solidity module
+   */
+  resolveTsModuleSync: (module: string, basedir: string) => string
+  /**
+   * Resolves cjs representation of the solidity module
+   */
+  resolveCjsModule: (module: string, basedir: string) => Promise<string>
+  /**
+   * Resolves cjs representation of the solidity module
+   */
+  resolveCjsModuleSync: (module: string, basedir: string) => string
+  /**
+   * Resolves the esm representation of the solidity module
+   */
+  resolveEsmModule: (module: string, basedir: string) => Promise<string>
+  /**
+   * Resolves the esm representation of the solidity module
+   */
+  resolveEsmModuleSync: (module: string, basedir: string) => string
+}
+
+export type Logger = {
+  info: (...messages: string[]) => void
+  error: (...message: string[]) => void
+  warn: (...message: string[]) => void
+  log: (...message: string[]) => void
+}
+/**
+ * Copied from rollup (kinda)
+ * @see https://rollupjs.org/plugin-development/#this-getmoduleinfo
+ */
+export interface ModuleInfo {
+  id: string // the id of the module, for convenience
+  rawCode: string | null // the source code of the module, `null` if external or not yet available
+  code: string | null // the code after transformed to correctly resolve remappings and node_module imports
+  importedIds: string[] // the module ids statically imported by this module
+  resolutions: ModuleInfo[] // how statically imported ids were resolved, for use with this.load
+}
+
+export type SolidityResolver = (
+  config: ResolvedConfig,
+  logger: Logger,
+) => {
+  /**
+   * The name of the plugin.
+   */
+  name: string
+  /**
+   * The configuration of the plugin.
+   */
+  config: ResolvedConfig
+  include?: string[]
+  exclude?: string[]
+  /**
+   * Resolves .d.ts representation of the solidity module
+   */
+  resolveDts: (module: string, basedir: string) => Promise<string>
+  /**
+   * Resolves .d.ts representation of the solidity module
+   */
+  resolveDtsSync: (module: string, basedir: string) => string
+  /**
+   * Resolves typescript representation of the solidity module
+   */
+  resolveTsModule: (module: string, basedir: string) => Promise<string>
+  /**
+   * Resolves typescript representation of the solidity module
+   */
+  resolveTsModuleSync: (module: string, basedir: string) => string
+  /**
+   * Resolves cjs representation of the solidity module
+   */
+  resolveCjsModule: (module: string, basedir: string) => Promise<string>
+  /**
+   * Resolves cjs representation of the solidity module
+   */
+  resolveCjsModuleSync: (module: string, basedir: string) => string
+  /**
+   * Resolves the esm representation of the solidity module
+   */
+  resolveEsmModule: (module: string, basedir: string) => Promise<string>
+  /**
+   * Resolves the esm representation of the solidity module
+   */
+  resolveEsmModuleSync: (module: string, basedir: string) => string
+}
