--- bundlers/bundler/src/types/FoundryToml.ts
@@ -1,8 +1,0 @@
-import { z } from 'zod'
-
-export const FoundryTomlSchema = z.object({
-	out: z.string().optional(),
-	artifacts: z.string().optional(),
-})
-
-export type FoundryToml = z.infer<typeof FoundryTomlSchema>
