--- bundlers/bundler/src/types/Logger.ts
@@ -1,6 +1,0 @@
-export type Logger = {
-	info: (...messages: string[]) => void
-	error: (...message: string[]) => void
-	warn: (...message: string[]) => void
-	log: (...message: string[]) => void
-}
