+++ bundlers/bundler/src/utils/formatPath.ts
@@ -1,0 +1,3 @@
+export const formatPath = (contractPath: string) => {
+	return contractPath.replace(/\\/g, '/')
+}
