+++ bundlers/bundler/src/utils/isImportLocal.ts
@@ -1,0 +1,3 @@
+export const isImportLocal = (importPath: string) => {
+	return importPath.startsWith('.')
+}
